requirements.txt
# Core ML (install a platform-appropriate PyTorch wheel first for best results)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and Hugging Face utilities
transformers>=4.30.0
huggingface-hub>=0.16.0

# Web app
streamlit>=1.28.0

# Image handling
Pillow>=9.5.0
opencv-python>=4.8.0

# Numerical and data processing
numpy>=1.24.0
pandas>=2.0.0

# Evaluation and analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0

# Notes:
# On Windows it's recommended to install a matching PyTorch wheel from https://pytorch.org/get-started/locally/
# Example (CPU-only) for PowerShell:
# python -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# Then install the remaining dependencies:
# pip install -r requirements.txt
#
# For GPU (CUDA 11.8):
# python -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install -r requirements.txt
